News_Aggregator/requirements.txt
ansicon==1.89.0
APScheduler==3.8.1
arrow==1.2.1
asgiref==3.4.1
beautifulsoup4==4.10.0
blessed==1.19.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-crontab==0.7.1
django-picklefield==3.0.1
django-q==1.3.9
django-restframework==0.0.1
djangorestframework==3.12.4
idna==3.3
jinxed==1.1.0
news==1.0
psycopg2==2.9.2
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wcwidth==0.2.5
